Improve code comments, fix int overflow for iterations, remove LFS
- Add meaningful comments to main.cpp, shapes.h, statistics.h, statistics.cpp
- Change iterations type from int to long long (prevents overflow >2B)
- Use std::stoll instead of std::stoi for CLI parsing
- Clear notebook outputs to reduce repo size
- Remove obsolete .gitattributes (LFS no longer needed)

+++ Datasaurus_Cpp/config.h
@@ -5,24 +5,24 @@
 
 namespace Config {
     // CLI parameter defaults
-    inline constexpr int DEFAULT_ITERATIONS = 1'000'000'000;
+    inline constexpr long long DEFAULT_ITERATIONS = 1'000'000'000;
     inline constexpr double DEFAULT_START_TEMP = 0.4;
     inline constexpr double DEFAULT_EPSILON = 0.003;
 
     // Internal constants (not exposed via CLI)
-    inline constexpr double END_TEMP = 0.00002;
-    inline constexpr double AMPLITUDE_FACTOR = 1.0;
-    inline constexpr double MIN_AMPLITUDE = 0.01;
+    inline constexpr double END_TEMP = 0.00002;      // Final temperature at which cooling stops
+    inline constexpr double AMPLITUDE_FACTOR = 1.0;   // Scales perturbation size relative to temperature
+    inline constexpr double MIN_AMPLITUDE = 0.01;     // Floor for perturbation size to keep points moving at low T
 
     // Computes cooling rate so temperature decays from startTemp to END_TEMP over n iterations
-    inline double computeCoolingRate(int iterations, double startTemp) {
+    inline double computeCoolingRate(long long iterations, double startTemp) {
         return std::pow(END_TEMP / startTemp, 1.0 / iterations);
     }
 }
 
 // Runtime configuration, populated from CLI arguments
 struct RuntimeConfig {
-    int totalIterations = Config::DEFAULT_ITERATIONS;
+    long long totalIterations = Config::DEFAULT_ITERATIONS;
     double startTemp = Config::DEFAULT_START_TEMP;
     double epsilon = Config::DEFAULT_EPSILON;
     double coolingRate = Config::computeCoolingRate(Config::DEFAULT_ITERATIONS, Config::DEFAULT_START_TEMP);
